# apps/hormuz-be/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Hormuz API",
    "version": "0.1.0"
  },
  "paths": {
    "/api/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "Health",
        "operationId": "health_api_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "type": "object",
                  "title": "Response Health Api Health Get"
                }
              }
            }
          }
        }
      }
    },
    "/api/scans/preview": {
      "post": {
        "tags": [
          "scan"
        ],
        "summary": "Preview Scan",
        "operationId": "preview_scan_api_scans_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/fixes/generate": {
      "post": {
        "tags": [
          "scan"
        ],
        "summary": "Generate Fixes",
        "operationId": "generate_fixes_api_fixes_generate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FixRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FixSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/scans/fixes": {
      "post": {
        "tags": [
          "scan"
        ],
        "summary": "Generate Scan Fixes",
        "operationId": "generate_scan_fixes_api_scans_fixes_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FixRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FixSummary"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "AgentFailure": {
        "properties": {
          "agent": {
            "type": "string",
            "title": "Agent"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "agent",
          "message"
        ],
        "title": "AgentFailure"
      },
      "Finding": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "violation_type": {
            "type": "string",
            "title": "Violation Type"
          },
          "agent": {
            "type": "string",
            "title": "Agent"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          },
          "file_path": {
            "type": "string",
            "title": "File Path"
          },
          "line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line"
          },
          "context": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snippet"
          },
          "regulations": {
            "items": {
              "$ref": "#/components/schemas/RegulationRef"
            },
            "type": "array",
            "title": "Regulations"
          },
          "regulation_warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Regulation Warning"
          },
          "recommendation": {
            "type": "string",
            "title": "Recommendation"
          },
          "remediation_hint": {
            "type": "string",
            "title": "Remediation Hint"
          }
        },
        "type": "object",
        "required": [
          "id",
          "violation_type",
          "agent",
          "category",
          "severity",
          "file_path",
          "title",
          "description",
          "recommendation",
          "remediation_hint"
        ],
        "title": "Finding"
      },
      "FixFailure": {
        "properties": {
          "finding_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finding Id"
          },
          "file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "FixFailure"
      },
      "FixOutputType": {
        "type": "string",
        "enum": [
          "local_diff",
          "github_pr"
        ],
        "title": "FixOutputType"
      },
      "FixPatch": {
        "properties": {
          "finding_id": {
            "type": "string",
            "title": "Finding Id"
          },
          "file_path": {
            "type": "string",
            "title": "File Path"
          },
          "diff": {
            "type": "string",
            "title": "Diff"
          },
          "patch_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patch Path"
          },
          "applied": {
            "type": "boolean",
            "title": "Applied",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "finding_id",
          "file_path",
          "diff"
        ],
        "title": "FixPatch"
      },
      "FixRequest": {
        "properties": {
          "repo_path": {
            "type": "string",
            "minLength": 1,
            "title": "Repo Path"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "title": "Findings"
          },
          "finding": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Finding"
              },
              {
                "type": "null"
              }
            ]
          },
          "apply": {
            "type": "boolean",
            "title": "Apply",
            "default": false
          },
          "create_pr": {
            "type": "boolean",
            "title": "Create Pr",
            "default": false
          },
          "rescan": {
            "type": "boolean",
            "title": "Rescan",
            "default": false
          },
          "patch_dir": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patch Dir"
          }
        },
        "type": "object",
        "required": [
          "repo_path"
        ],
        "title": "FixRequest"
      },
      "FixSummary": {
        "properties": {
          "output_type": {
            "$ref": "#/components/schemas/FixOutputType"
          },
          "pr_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pr Url"
          },
          "patch_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Patch Path"
          },
          "diff": {
            "type": "string",
            "title": "Diff"
          },
          "patches": {
            "items": {
              "$ref": "#/components/schemas/FixPatch"
            },
            "type": "array",
            "title": "Patches"
          },
          "failures": {
            "items": {
              "$ref": "#/components/schemas/FixFailure"
            },
            "type": "array",
            "title": "Failures"
          },
          "applied": {
            "type": "boolean",
            "title": "Applied",
            "default": false
          },
          "rescan_summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ScanSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "output_type",
          "diff",
          "patches"
        ],
        "title": "FixSummary"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RegulationRef": {
        "properties": {
          "framework": {
            "type": "string",
            "enum": [
              "GDPR",
              "APP"
            ],
            "title": "Framework"
          },
          "clause": {
            "type": "string",
            "title": "Clause"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          },
          "requirement": {
            "type": "string",
            "title": "Requirement"
          },
          "max_penalty": {
            "type": "string",
            "title": "Max Penalty"
          },
          "severity": {
            "$ref": "#/components/schemas/Severity"
          }
        },
        "type": "object",
        "required": [
          "framework",
          "clause",
          "title",
          "summary",
          "requirement",
          "max_penalty",
          "severity"
        ],
        "title": "RegulationRef"
      },
      "ScanRequest": {
        "properties": {
          "repo_path": {
            "type": "string",
            "minLength": 1,
            "title": "Repo Path"
          }
        },
        "type": "object",
        "required": [
          "repo_path"
        ],
        "title": "ScanRequest"
      },
      "ScanStatus": {
        "type": "string",
        "enum": [
          "complete",
          "partial",
          "failed"
        ],
        "title": "ScanStatus"
      },
      "ScanSummary": {
        "properties": {
          "scan_status": {
            "$ref": "#/components/schemas/ScanStatus"
          },
          "score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "total_findings": {
            "type": "integer",
            "title": "Total Findings"
          },
          "counts_by_severity": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/Severity"
            },
            "type": "object",
            "title": "Counts By Severity"
          },
          "counts_by_agent": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Counts By Agent"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "title": "Findings"
          },
          "failed_agents": {
            "items": {
              "$ref": "#/components/schemas/AgentFailure"
            },
            "type": "array",
            "title": "Failed Agents"
          }
        },
        "type": "object",
        "required": [
          "scan_status",
          "score",
          "total_findings",
          "counts_by_severity",
          "counts_by_agent",
          "findings",
          "failed_agents"
        ],
        "title": "ScanSummary"
      },
      "Severity": {
        "type": "string",
        "enum": [
          "critical",
          "high",
          "medium",
          "low"
        ],
        "title": "Severity"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}